+++ 4180Project_COMConnector/4180Project_COMConnector.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.5.33530.505
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "4180_ProjectCOM", "4180_ProjectCOM\4180_ProjectCOM.csproj", "{A30D6065-C33E-48C1-908C-0441F5156CA9}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "4180Project_COMConnector", "4180Project_COMConnector\4180Project_COMConnector.csproj", "{6EF4FAA1-4345-40EE-8521-F66DFBC88F3D}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,15 +11,15 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{A30D6065-C33E-48C1-908C-0441F5156CA9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A30D6065-C33E-48C1-908C-0441F5156CA9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A30D6065-C33E-48C1-908C-0441F5156CA9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A30D6065-C33E-48C1-908C-0441F5156CA9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6EF4FAA1-4345-40EE-8521-F66DFBC88F3D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6EF4FAA1-4345-40EE-8521-F66DFBC88F3D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6EF4FAA1-4345-40EE-8521-F66DFBC88F3D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6EF4FAA1-4345-40EE-8521-F66DFBC88F3D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {D04A1747-2847-4483-9AFC-ADE6FA54C3D0}
+		SolutionGuid = {6C489E02-86AB-44E2-93C0-BBB4BBDF6983}
 	EndGlobalSection
 EndGlobal
