--- 4180_ProjectCOM/4180_ProjectCOM/Form1.cs
@@ -1,10 +1,0 @@
-namespace _4180_ProjectCOM
-{
-    public partial class Form1 : Form
-    {
-        public Form1()
-        {
-            InitializeComponent();
-        }
-    }
-}
